Move hearing detail functions into their own files
