+++ packages/core/phase3/lib/hearingDetails/preProcessOffenceReasonSequence.ts
@@ -1,0 +1,4 @@
+import type { Offence } from "../../../types/AnnotatedHearingOutcome"
+
+export const preProcessOffenceReasonSequence = (offence: Offence): string =>
+  offence.CriminalProsecutionReference.OffenceReasonSequence?.padStart(3, "0") ?? ""
