+++ packages/core/phase3/lib/pncUpdateRequestGenerators/disposalUpdatedGenerator.ts
@@ -6,7 +6,7 @@ import formatDateSpecifiedInResult from "../../../phase2/lib/createPncDisposalsF
 import { PncOperation } from "../../../types/PncOperation"
 import generateBasePncUpdateRequest from "../generateBasePncUpdateRequest"
 import getPncCourtCode from "../getPncCourtCode"
-import { generateHearingsAdjudicationsAndDisposals } from "../hearingDetails"
+import { generateHearingsAdjudicationsAndDisposals } from "../hearingDetails/generateHearingsAdjudicationsAndDisposals"
 import { preProcessCourtCaseReferenceNumber } from "./normalDisposalGenerator"
 
 const DISPOSAL_UPDATED_HEARING_TYPE = "V"
